feat: add support to Arch and Manjaro linux

+++ tools/alacritty/install.sh
@@ -6,19 +6,26 @@ TEMPORARY_DIRECORY="$(mktemp -d)/alacritty"
 SCRIPT_PATH="$(dirname -- "${BASH_SOURCE[0]}")"
 SETTINGS_DIRECTORY="${SCRIPT_PATH}/settings"
 
-sudo apt-get install --assume-yes cmake pkg-config libfreetype6-dev libfontconfig1-dev libxcb-xfixes0-dev \
-    libxkbcommon-dev autotools-dev automake libncurses-dev
+source "${SCRIPT_PATH}/../../common.sh"
 
-git clone --depth=1 https://github.com/alacritty/alacritty.git "${TEMPORARY_DIRECORY}"
-rm -rf ~/.config/alacritty/
+os_name=$(get_os)
+if [[ ${os_name} == Ubuntu* ]]; then
+    sudo apt-get install --assume-yes cmake pkg-config libfreetype6-dev libfontconfig1-dev libxcb-xfixes0-dev \
+        libxkbcommon-dev autotools-dev automake libncurses-dev
 
-pushd "${TEMPORARY_DIRECORY}"
-    cargo build --release
-    sudo rm -f /usr/local/bin/alacritty
-    sudo cp target/release/alacritty /usr/local/bin
-popd
+    git clone --depth=1 https://github.com/alacritty/alacritty.git "${TEMPORARY_DIRECORY}"
+    rm -rf ~/.config/alacritty/
 
-rm -rf "${TEMPORARY_DIRECORY}"
+    pushd "${TEMPORARY_DIRECORY}"
+        cargo build --release
+        sudo rm -f /usr/local/bin/alacritty
+        sudo cp target/release/alacritty /usr/local/bin
+    popd
+
+    rm -rf "${TEMPORARY_DIRECORY}"
+elif [[ ${os_name} == Manjaro* ]]; then
+    sudo pacman -S alacritty
+fi
 
 mkdir -p ~/.config/alacritty/
 cp -r "${SETTINGS_DIRECTORY}"/* ~/.config/alacritty/
