+++ tools/lazygit/install.sh
@@ -2,15 +2,24 @@
 
 set -xeu
 
-VERSION="0.40.2"
-TEMPORARY_DIRECORY="$(mktemp -d)"
+SCRIPT_PATH="$(dirname -- "${BASH_SOURCE[0]}")"
 
-wget https://github.com/jesseduffield/lazygit/releases/download/v${VERSION}/lazygit_${VERSION}_Linux_x86_64.tar.gz \
-    -P "${TEMPORARY_DIRECORY}"
+source "${SCRIPT_PATH}/../../common.sh"
 
-pushd "${TEMPORARY_DIRECORY}"
-    tar -xf lazygit_${VERSION}_Linux_x86_64.tar.gz
-    sudo cp lazygit /usr/bin
-popd
+os_name=$(get_os)
+if [[ ${os_name} == Ubuntu* ]]; then
+    VERSION="0.40.2"
+    TEMPORARY_DIRECORY="$(mktemp -d)"
 
-rm -rf "${TEMPORARY_DIRECORY}"
+    wget https://github.com/jesseduffield/lazygit/releases/download/v${VERSION}/lazygit_${VERSION}_Linux_x86_64.tar.gz \
+        -P "${TEMPORARY_DIRECORY}"
+
+    pushd "${TEMPORARY_DIRECORY}"
+        tar -xf lazygit_${VERSION}_Linux_x86_64.tar.gz
+        sudo cp lazygit /usr/bin
+    popd
+
+    rm -rf "${TEMPORARY_DIRECORY}"
+elif [[ ${os_name} == Manjaro* ]]; then
+    sudo pacman -S lazygit
+fi
