+++ tools/ranger/install.sh
@@ -6,17 +6,25 @@ TEMPORARY_DIRECORY="$(mktemp -d)/ranger"
 SCRIPT_PATH="$(dirname -- "${BASH_SOURCE[0]}")"
 SETTINGS_DIRECTORY="${SCRIPT_PATH}/settings"
 
-sudo apt install --assume-yes dialog
-
-git clone --depth=1 https://github.com/ranger/ranger.git "${TEMPORARY_DIRECORY}"
 sudo rm -rf ~/.config/ranger/
 sudo rm -rf /root/.config/ranger
 
-pushd "${TEMPORARY_DIRECORY}"
-    sudo make install
-popd
+source "${SCRIPT_PATH}/../../common.sh"
 
-sudo rm -rf "${TEMPORARY_DIRECORY}"
+os_name=$(get_os)
+if [[ ${os_name} == Ubuntu* ]]; then
+    sudo apt install --assume-yes dialog
+
+    git clone --depth=1 https://github.com/ranger/ranger.git "${TEMPORARY_DIRECORY}"
+
+    pushd "${TEMPORARY_DIRECORY}"
+        sudo make install
+    popd
+
+    sudo rm -rf "${TEMPORARY_DIRECORY}"
+elif [[ ${os_name} == Manjaro* ]]; then
+    sudo pacman -S ranger
+fi
 
 ranger --copy-config all
 cp -r "${SETTINGS_DIRECTORY}"/* ~/.config/ranger/
